requirements.txt
aiosqlite
asyncpg
tortoise-orm
pyyaml
